baseline_1/requirements.txt
timm==0.6.7
pillow==9.2.0
matplotlib==3.5.3
torchprofile==0.0.4
seaborn
scikit-learn
numpy==1.26.4
